Split Last.fm/Libre.fm tests
