+++ tests/test_pylast.py
@@ -2,11 +2,12 @@
 """
 Integration (not unit) tests for pylast.py
 """
-from flaky import flaky
 import os
-import pytest
 import time
 import unittest
+
+import pytest
+from flaky import flaky
 
 import pylast
 
@@ -1393,26 +1394,6 @@ class TestPyLastUser(PyLastTestCase):
 
         # Assert
         self.assertEqual(mbid, None)
-
-
-@flaky(max_runs=5, min_passes=1)
-class TestPyLastWithLibreFm(unittest.TestCase):
-    """Own class for Libre.fm because we don't need the Last.fm setUp"""
-
-    def test_libre_fm(self):
-        # Arrange
-        secrets = load_secrets()
-        username = secrets["username"]
-        password_hash = secrets["password_hash"]
-
-        # Act
-        network = pylast.LibreFMNetwork(
-            password_hash=password_hash, username=username)
-        artist = network.get_artist("Radiohead")
-        name = artist.get_name()
-
-        # Assert
-        self.assertEqual(name, "Radiohead")
 
 
 if __name__ == '__main__':
